+++ src/device.rs
@@ -191,158 +191,3 @@ impl DeviceFlow {
         Token::from_json(&body)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use hyper_rustls::HttpsConnector;
-    use std::pin::Pin;
-
-    use super::*;
-
-    #[tokio::test]
-    async fn test_device_end2end() {
-        #[derive(Clone)]
-        struct FD;
-        impl DeviceFlowDelegate for FD {
-            fn present_user_code<'a>(
-                &'a self,
-                pi: &'a DeviceAuthResponse,
-            ) -> Pin<Box<dyn Future<Output = ()> + 'a + Send>> {
-                assert_eq!("https://example.com/verify", pi.verification_uri);
-                Box::pin(futures::future::ready(()))
-            }
-        }
-
-        let server_url = mockito::server_url();
-        let app_secret: ApplicationSecret = crate::parse_json!({
-            "client_id": "902216714886-k2v9uei3p1dk6h686jbsn9mo96tnbvto.apps.googleusercontent.com",
-            "project_id": "yup-test-243420",
-            "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-            "token_uri": format!("{}/token", server_url),
-            "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-            "client_secret": "iuMPN6Ne1PD7cos29Tk9rlqH",
-            "redirect_uris": ["urn:ietf:wg:oauth:2.0:oob","http://localhost"],
-        });
-        let device_code_url = format!("{}/code", server_url);
-
-        let https = HttpsConnector::new();
-        let client = hyper::Client::builder()
-            .keep_alive(false)
-            .build::<_, hyper::Body>(https);
-
-        let flow = DeviceFlow {
-            app_secret,
-            device_code_url: device_code_url.into(),
-            flow_delegate: Box::new(FD),
-            grant_type: GOOGLE_GRANT_TYPE.into(),
-        };
-
-        // Successful path
-        {
-            let code_response = serde_json::json!({
-                "device_code": "devicecode",
-                "user_code": "usercode",
-                "verification_url": "https://example.com/verify",
-                "expires_in": 1234567,
-                "interval": 1
-            });
-            let _m = mockito::mock("POST", "/code")
-                .match_body(mockito::Matcher::Regex(
-                    ".*client_id=902216714886.*".to_string(),
-                ))
-                .with_status(200)
-                .with_body(code_response.to_string())
-                .create();
-            let token_response = serde_json::json!({
-                "access_token": "accesstoken",
-                "refresh_token": "refreshtoken",
-                "token_type": "Bearer",
-                "expires_in": 1234567
-            });
-            let _m = mockito::mock("POST", "/token")
-                .match_body(mockito::Matcher::Regex(
-                    ".*client_secret=iuMPN6Ne1PD7cos29Tk9rlqH&code=devicecode.*".to_string(),
-                ))
-                .with_status(200)
-                .with_body(token_response.to_string())
-                .create();
-
-            let token = flow
-                .token(&client, &["https://www.googleapis.com/scope/1"])
-                .await
-                .expect("token failed");
-            assert_eq!("accesstoken", token.access_token);
-            _m.assert();
-        }
-
-        // Code is not delivered.
-        {
-            let code_response = serde_json::json!({
-                "error": "invalid_client_id",
-                "error_description": "description"
-            });
-            let _m = mockito::mock("POST", "/code")
-                .match_body(mockito::Matcher::Regex(
-                    ".*client_id=902216714886.*".to_string(),
-                ))
-                .with_status(400)
-                .with_body(code_response.to_string())
-                .create();
-            let token_response = serde_json::json!({
-                "access_token": "accesstoken",
-                "refresh_token": "refreshtoken",
-                "token_type": "Bearer",
-                "expires_in": 1234567
-            });
-            let _m = mockito::mock("POST", "/token")
-                .match_body(mockito::Matcher::Regex(
-                    ".*client_secret=iuMPN6Ne1PD7cos29Tk9rlqH&code=devicecode.*".to_string(),
-                ))
-                .with_status(200)
-                .with_body(token_response.to_string())
-                .expect(0) // Never called!
-                .create();
-
-            let res = flow
-                .token(&client, &["https://www.googleapis.com/scope/1"])
-                .await;
-            assert!(res.is_err());
-            assert!(format!("{}", res.unwrap_err()).contains("invalid_client_id"));
-            _m.assert();
-        }
-
-        // Token is not delivered.
-        {
-            let code_response = serde_json::json!({
-                "device_code": "devicecode",
-                "user_code": "usercode",
-                "verification_url": "https://example.com/verify",
-                "expires_in": 1234567,
-                "interval": 1
-            });
-            let _m = mockito::mock("POST", "/code")
-                .match_body(mockito::Matcher::Regex(
-                    ".*client_id=902216714886.*".to_string(),
-                ))
-                .with_status(200)
-                .with_body(code_response.to_string())
-                .create();
-            let token_response = serde_json::json!({"error": "access_denied"});
-            let _m = mockito::mock("POST", "/token")
-                .match_body(mockito::Matcher::Regex(
-                    ".*client_secret=iuMPN6Ne1PD7cos29Tk9rlqH&code=devicecode.*".to_string(),
-                ))
-                .with_status(400)
-                .with_body(token_response.to_string())
-                .expect(1)
-                .create();
-
-            let res = flow
-                .token(&client, &["https://www.googleapis.com/scope/1"])
-                .await;
-            assert!(res.is_err());
-            assert!(format!("{}", res.unwrap_err()).contains("access_denied"));
-            _m.assert();
-        }
-    }
-}
