+++ src/installed.rs
@@ -354,170 +354,8 @@ mod installed_flow_server {
 
 #[cfg(test)]
 mod tests {
-    use std::str::FromStr;
-
-    use hyper::client::connect::HttpConnector;
-    use hyper::Uri;
-    use hyper_rustls::HttpsConnector;
-    use mockito::mock;
-
     use super::*;
-    use crate::authenticator_delegate::InstalledFlowDelegate;
-
-    #[tokio::test]
-    async fn test_end2end() {
-        #[derive(Clone)]
-        struct FD(
-            String,
-            hyper::Client<HttpsConnector<HttpConnector>, hyper::Body>,
-        );
-        impl InstalledFlowDelegate for FD {
-            /// Depending on need_code, return the pre-set code or send the code to the server at
-            /// the redirect_uri given in the url.
-            fn present_user_url<'a>(
-                &'a self,
-                url: &'a str,
-                need_code: bool,
-            ) -> Pin<Box<dyn Future<Output = Result<String, String>> + Send + 'a>> {
-                Box::pin(async move {
-                    if need_code {
-                        Ok(self.0.clone())
-                    } else {
-                        // Parse presented url to obtain redirect_uri with location of local
-                        // code-accepting server.
-                        let uri = Uri::from_str(url.as_ref()).unwrap();
-                        let query = uri.query().unwrap();
-                        let parsed = form_urlencoded::parse(query.as_bytes()).into_owned();
-                        let mut rduri = None;
-                        for (k, v) in parsed {
-                            if k == "redirect_uri" {
-                                rduri = Some(v);
-                                break;
-                            }
-                        }
-                        if rduri.is_none() {
-                            return Err("no redirect_uri!".into());
-                        }
-                        let mut rduri = rduri.unwrap();
-                        rduri.push_str(&format!("?code={}", self.0));
-                        let rduri = Uri::from_str(rduri.as_ref()).unwrap();
-                        // Hit server.
-                        self.1
-                            .get(rduri)
-                            .await
-                            .map_err(|e| e.to_string())
-                            .map(|_| "".to_string())
-                    }
-                })
-            }
-        }
-
-        let server_url = mockito::server_url();
-        let app_secret: ApplicationSecret = crate::parse_json!({
-            "client_id": "902216714886-k2v9uei3p1dk6h686jbsn9mo96tnbvto.apps.googleusercontent.com",
-            "project_id": "yup-test-243420",
-            "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-            "token_uri": format!("{}/token", server_url),
-            "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-            "client_secret": "iuMPN6Ne1PD7cos29Tk9rlqH",
-            "redirect_uris": ["urn:ietf:wg:oauth:2.0:oob","http://localhost"]
-        });
-
-        let https = HttpsConnector::new();
-        let client = hyper::Client::builder()
-            .keep_alive(false)
-            .build::<_, hyper::Body>(https);
-
-        let fd = FD("authorizationcode".to_string(), client.clone());
-        let inf = InstalledFlow {
-            app_secret: app_secret.clone(),
-            method: InstalledFlowReturnMethod::Interactive,
-            flow_delegate: Box::new(fd),
-        };
-
-        // Successful path.
-        {
-            let _m = mock("POST", "/token")
-                .match_body(mockito::Matcher::Regex(
-                    ".*code=authorizationcode.*client_id=9022167.*".to_string(),
-                ))
-                .with_body(
-                    serde_json::json!({
-                        "access_token": "accesstoken",
-                        "refresh_token": "refreshtoken",
-                        "token_type": "Bearer",
-                        "expires_in": 12345678
-                    })
-                    .to_string(),
-                )
-                .expect(1)
-                .create();
-
-            let tok = inf
-                .token(&client, &["https://googleapis.com/some/scope"])
-                .await
-                .expect("failed to get token");
-            assert_eq!("accesstoken", tok.access_token);
-            assert_eq!("refreshtoken", tok.refresh_token.unwrap());
-            assert_eq!("Bearer", tok.token_type);
-            _m.assert();
-        }
-
-        // Successful path with HTTP redirect.
-        {
-            let inf = InstalledFlow {
-                app_secret: app_secret.clone(),
-                method: InstalledFlowReturnMethod::HTTPRedirect,
-                flow_delegate: Box::new(FD(
-                    "authorizationcodefromlocalserver".to_string(),
-                    client.clone(),
-                )),
-            };
-            let _m = mock("POST", "/token")
-                .match_body(mockito::Matcher::Regex(
-                    ".*code=authorizationcodefromlocalserver.*client_id=9022167.*".to_string(),
-                ))
-                .with_body(
-                    serde_json::json!({
-                        "access_token": "accesstoken",
-                        "refresh_token": "refreshtoken",
-                        "token_type": "Bearer",
-                        "expires_in": 12345678
-                    })
-                    .to_string(),
-                )
-                .expect(1)
-                .create();
-
-            let tok = inf
-                .token(&client, &["https://googleapis.com/some/scope"])
-                .await
-                .expect("failed to get token");
-            assert_eq!("accesstoken", tok.access_token);
-            assert_eq!("refreshtoken", tok.refresh_token.unwrap());
-            assert_eq!("Bearer", tok.token_type);
-            _m.assert();
-        }
-
-        // Error from server.
-        {
-            let _m = mock("POST", "/token")
-                .match_body(mockito::Matcher::Regex(
-                    ".*code=authorizationcode.*client_id=9022167.*".to_string(),
-                ))
-                .with_status(400)
-                .with_body(serde_json::json!({"error": "invalid_code"}).to_string())
-                .expect(1)
-                .create();
-
-            let tokr = inf
-                .token(&client, &["https://googleapis.com/some/scope"])
-                .await;
-            assert!(tokr.is_err());
-            assert!(format!("{}", tokr.unwrap_err()).contains("invalid_code"));
-            _m.assert();
-        }
-    }
+    use hyper::Uri;
 
     #[test]
     fn test_request_url_builder() {
