+++ src/refresh.rs
@@ -57,64 +57,3 @@ impl RefreshFlow {
         Ok(token)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::helper;
-
-    use hyper_rustls::HttpsConnector;
-
-    #[tokio::test]
-    async fn test_refresh_end2end() {
-        let server_url = mockito::server_url();
-
-        let app_secret = r#"{"installed":{"client_id":"902216714886-k2v9uei3p1dk6h686jbsn9mo96tnbvto.apps.googleusercontent.com","project_id":"yup-test-243420","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"iuMPN6Ne1PD7cos29Tk9rlqH","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}"#;
-        let mut app_secret = helper::parse_application_secret(app_secret).unwrap();
-        app_secret.token_uri = format!("{}/token", server_url);
-        let refresh_token = "my-refresh-token";
-
-        let https = HttpsConnector::new();
-        let client = hyper::Client::builder()
-            .keep_alive(false)
-            .build::<_, hyper::Body>(https);
-
-        // Success
-        {
-            let _m = mockito::mock("POST", "/token")
-                .match_body(
-                    mockito::Matcher::Regex(".*client_id=902216714886-k2v9uei3p1dk6h686jbsn9mo96tnbvto.apps.googleusercontent.com.*refresh_token=my-refresh-token.*".to_string()))
-                .with_status(200)
-                .with_body(r#"{"access_token": "new-access-token", "token_type": "Bearer", "expires_in": 1234567}"#)
-                .create();
-            let token = RefreshFlow::refresh_token(&client, &app_secret, refresh_token)
-                .await
-                .expect("token failed");
-            assert_eq!("new-access-token", token.access_token);
-            assert_eq!("Bearer", token.token_type);
-            _m.assert();
-        }
-
-        // Refresh error.
-        {
-            let _m = mockito::mock("POST", "/token")
-                .match_body(
-                    mockito::Matcher::Regex(".*client_id=902216714886-k2v9uei3p1dk6h686jbsn9mo96tnbvto.apps.googleusercontent.com.*refresh_token=my-refresh-token.*".to_string()))
-                .with_status(400)
-                .with_body(r#"{"error": "invalid_request"}"#)
-                .create();
-
-            let rr = RefreshFlow::refresh_token(&client, &app_secret, refresh_token).await;
-            match rr {
-                Err(Error::AuthError(auth_error)) => {
-                    assert_eq!(
-                        auth_error.error,
-                        crate::error::AuthErrorCode::InvalidRequest
-                    );
-                }
-                _ => panic!(format!("unexpected RefreshResult {:?}", rr)),
-            }
-            _m.assert();
-        }
-    }
-}
